+++ src/errors.rs
@@ -39,6 +39,15 @@ pub enum ErrorKind {
     CallFilter(String),
     /// An error occured while executing a test.
     CallTest(String),
+    /// An IO error occured
+    Io(std::io::ErrorKind),
+    /// UTF-8 conversion error
+    ///
+    /// This should not occur unless invalid UTF8 chars are rendered
+    Utf8Conversion {
+        /// The context that indicates where the error occurs in the rendering process
+        context: String,
+    },
     /// This enum may grow additional variants, so this makes sure clients
     /// don't count on exhaustive matching. (Otherwise, adding a new variant
     /// could break existing code.)
@@ -79,6 +88,12 @@ impl fmt::Display for Error {
             ErrorKind::CallFunction(ref name) => write!(f, "Function call '{}' failed", name),
             ErrorKind::CallFilter(ref name) => write!(f, "Filter call '{}' failed", name),
             ErrorKind::CallTest(ref name) => write!(f, "Test call '{}' failed", name),
+            ErrorKind::Io(ref io_error) => {
+                write!(f, "Io error while writing rendered value to output: {:?}", io_error)
+            }
+            ErrorKind::Utf8Conversion { ref context } => {
+                write!(f, "UTF-8 conversion error occured while rendering template: {}", context)
+            }
             ErrorKind::__Nonexhaustive => write!(f, "Nonexhaustive"),
         }
     }
@@ -173,8 +188,23 @@ impl Error {
     pub fn invalid_macro_def(name: impl ToString) -> Self {
         Self { kind: ErrorKind::InvalidMacroDefinition(name.to_string()), source: None }
     }
+
+    /// Creates an IO error
+    pub fn io_error(error: std::io::Error) -> Self {
+        Self { kind: ErrorKind::Io(error.kind()), source: Some(Box::new(error)) }
+    }
+
+    /// Creates an utf8 conversion error
+    pub fn utf8_conversion_error(error: std::string::FromUtf8Error, context: String) -> Self {
+        Self { kind: ErrorKind::Utf8Conversion { context }, source: Some(Box::new(error)) }
+    }
 }
 
+impl From<std::io::Error> for Error {
+    fn from(error: std::io::Error) -> Self {
+        Self::io_error(error)
+    }
+}
 impl From<&str> for Error {
     fn from(e: &str) -> Self {
         Self::msg(e)